+++ realm-annotations/src/main/java/io/realm/annotations/RealmClass.java
@@ -37,6 +37,16 @@ public @interface RealmClass {
      * {@link RealmModule#classNamingPolicy()}.
      *
      * @see io.realm.annotations.RealmNamingPolicy for more information about what setting the name means.
+     * @see #name()
+     */
+    String value() default "";
+
+    /**
+     * Manually set the internal name used by Realm for this class. If this class is part of
+     * any modules, this will also override any name policy set using
+     * {@link RealmModule#classNamingPolicy()}.
+     *
+     * @see io.realm.annotations.RealmNamingPolicy for more information about what setting the name means.
      */
     String name() default "";