+++ realm/realm-library/src/syncTestUtils/java/io/realm/SyncTestUtils.java
@@ -14,11 +14,15 @@
  * limitations under the License.
  */
 
-package io.realm.util;
+package io.realm;
+
+import android.support.test.InstrumentationRegistry;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.io.File;
+import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.UUID;
@@ -33,6 +37,9 @@ import io.realm.SyncUser;
 import io.realm.UserStore;
 import io.realm.internal.network.AuthenticateResponse;
 import io.realm.internal.objectserver.Token;
+import io.realm.log.LogLevel;
+import io.realm.log.RealmLog;
+import io.realm.objectserver.utils.UserFactory;
 
 public class SyncTestUtils {
 
@@ -41,6 +48,7 @@ public class SyncTestUtils {
 
     private final static Method SYNC_MANAGER_GET_USER_STORE_METHOD;
     private final static Method SYNC_USER_GET_ACCESS_TOKEN_METHOD;
+    private static int originalLogLevel; // Should only be modified by prepareEnvironmentForTest and restoreEnvironmentAfterTest
     static {
         try {
             SYNC_MANAGER_GET_USER_STORE_METHOD = SyncManager.class.getDeclaredMethod("getUserStore");
@@ -49,6 +57,55 @@ public class SyncTestUtils {
             SYNC_USER_GET_ACCESS_TOKEN_METHOD.setAccessible(true);
         } catch (NoSuchMethodException e) {
             throw new AssertionError(e);
+        }
+    }
+
+    public static void prepareEnvironmentForTest() throws IOException {
+        deleteRosFiles();
+        if (BaseRealm.applicationContext != null) {
+            // Realm was already initialized. Reset all internal state
+            // in order to be able fully re-initialize.
+
+            // This will set the 'm_metadata_manager' in 'sync_manager.cpp' to be 'null'
+            // causing the SyncUser to remain in memory.
+            // They're actually not persisted into disk.
+            // move this call to 'tearDown' to clean in-memory & on-disk users
+            // once https://github.com/realm/realm-object-store/issues/207 is resolved
+            SyncManager.reset();
+            BaseRealm.applicationContext = null; // Required for Realm.init() to work
+        }
+        Realm.init(InstrumentationRegistry.getTargetContext());
+        originalLogLevel = RealmLog.getLevel();
+        RealmLog.setLevel(LogLevel.DEBUG);
+    }
+
+    /**
+     * Tries to restore the environment as best as possible after a test.
+     */
+    public static void restoreEnvironmentAfterTest() {
+        // Block until all users are logged out
+        UserFactory.logoutAllUsers();
+
+        // Reset log level
+        RealmLog.setLevel(originalLogLevel);
+    }
+
+    // Cleanup filesystem to make sure nothing lives for the next test.
+    // Failing to do so might lead to DIVERGENT_HISTORY errors being thrown if Realms from
+    // previous tests are being accessed.
+    private static void deleteRosFiles() throws IOException {
+        File rosFiles = new File(InstrumentationRegistry.getContext().getFilesDir(),"realm-object-server");
+        deleteFile(rosFiles);
+    }
+
+    private static void deleteFile(File file) throws IOException {
+        if (file.isDirectory()) {
+            for (File c : file.listFiles()) {
+                deleteFile(c);
+            }
+        }
+        if (!file.delete()) {
+            throw new IllegalStateException("Failed to delete file or directory: " + file.getAbsolutePath());
         }
     }