+++ CHANGELOG.md
@@ -1,22 +1,37 @@
-## 5.5.0 (YYYY-MM-DD)
+## 5.6.0 (YYYY-MM-DD)
+
+### Enhancements
+
+* `@RealmClass("name")` and `@RealmField("name")` can now be used as a shorthand for defining custom name mappings (#6145).
+* Added support for `RealmQuery.limit(long limit)` (#544).
+
+
+## 5.5.0 (2018-08-31)
 
 ### Enhancements
 
 * [ObjectServer] Added `ConnectionState` enum describing the states a connection can be in.
-* [ObjectServer] Added `SyncSession.isConnected()`.
+* [ObjectServer] Added `SyncSession.isConnected()` and `SyncSession.getConnectionState()`.
 * [ObjectServer] Added support for observing connection changes for a session using `SyncSession.addConnectionChangeListener()` and `SyncSession.removeConnectionChangeListener()`.
 * [ObjectServer] Added Kotlin extension property `Realm.syncSession` for synchronized Realms.
 * [ObjectServer] Added Kotlin extension method `Realm.classPermissions<RealmModel>()`.
 * [ObjectServer] Added support for starting and stopping synchronization using `SyncSession.start()` and `SyncSession.stop()` (#6135).
-* Added support for `RealmQuery.limit(long limit)` (#544).
-
+* [ObjectServer] Added API's for making it easier to work with network proxies (#6163): 
+  * `SyncManager.setAuthorizationHeaderName(String headerName)`
+  * `SyncManager.setAuthorizationHeaderName(String headerName, String host)`
+  * `SyncManager.addCustomRequestHeader(String headerName, String headerValue)`
+  * `SyncManager.addCustomRequestHeader(String headerName, String headerValue, String host)`
+  * `SyncManager.addCustomRequestHeaders(Map<String, String> headers)`
+  * `SyncManager.addCustomRequestHeaders(Map<String, String> headers, String host)`
+  * `SyncConfiguration.Builder.urlPrefix(String prefix)`
+ 
 ### Bug Fixes
 
 * Methods and classes requiring synchronized Realms have been removed from the standard AAR package. They are now only visible when enabling synchronized Realms in Gradle. The methods and classes will still be visible in the source files and docs, but annotated with `@ObjectServer` (#5799).
 
 ### Internal
 
-* Updated to Realm Sync 3.9.3
+* Updated to Realm Sync 3.9.4
 * Updated to Realm Core 5.8.0
 * Updated to Object Store commit: b0fc2814d9e6061ce5ba1da887aab6cfba4755ca